feat: add fuzzy search functionality
This commit adds the Fuse.js library as a dependency and uses it to perform fuzzy searches on the list of masked emails. A new SearchBar component is also added to the TopComponent, which allows users to input their search queries. The search results are then displayed in the MaskedEmailListPane component.