feat: add Firefox support to build process
This commit adds support for building the extension in Firefox by adding a new environment variable `BROWSER` and updating the `build:production:firefox` script. The `manifest.ts` file was also updated to use this new variable instead of the old one. Additionally, the `update-version.js` script was modified to update both Chrome and Firefox manifest files with sanitized version numbers.