feat(dropdown): Remember selected filter
The filter drop down currently always uses All by default.
This default can be annoying if you have a deleted masked emails
that sort near the top.

This changes FilterEmailsDropdown to remember the selected filter
between invocations, so the default changes to what you selected last
time.

Note that it remembers the value only when changed with the drop down
and not when you create a new masked email.